Include Google Data API and my own Localization manager
Yeah - I did wrote a "localization manager" because I just hate the
.NET's one. (With satellite assemblies)
This one works with embedded txt-files. I will do some
changes/optimations to it in the future but for now that one will do it.

I don't think I have to say anything to the Google Data API. I guess
that's clear. :D