Add new class RestoreData
This is an awesome class.
It should save all important data which is currently in memory. So it
would be redundant IF you couldn't serialize it. But you can so you can
write the whole object to a file and have your memory backup! ;-)
If the program crashes you can "restore the data" and that's it. :D